# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
...
        "/quizzes": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get Quizzes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quizzes"
                ],
                "responses": {
                    "200": {
                        "description": "success get quizzes",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helper.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.Quiz"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorResponse"
                        }
                    },
                    "408": {
                        "description": "Request Timeout",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/quizzes/check": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Check Answer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quizzes"
                ],
                "parameters": [
                    {
                        "description": "answer",
                        "name": "answer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.AnswerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success check answer",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helper.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.AnswerResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorResponse"
                        }
                    },
                    "408": {
                        "description": "Request Timeout",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.AnswerRequest": {
            "type": "object",
            "properties": {
                "quiz_id": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "user_answer": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "domain.AnswerResponse": {
            "type": "object",
            "properties": {
                "correct_answer": {
                    "type": "integer"
                },
                "user_experience": {
                    "type": "integer"
                },
                "user_level": {
                    "type": "integer"
                },
                "user_tier": {
                    "type": "string"
                }
            }
        },
...
        "domain.Quiz": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string"
                },
                "difficulty": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_link": {
                    "type": "string"
                },
                "option_a": {
                    "type": "string"
                },
                "option_b": {
                    "type": "string"
                },
                "option_c": {
                    "type": "string"
                },
                "option_d": {
                    "type": "string"
                },
                "question": {
                    "type": "string"
                },
                "user_quiz": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.UserAnswer"
                    }
                }
            }
        },
...
        "domain.UserAnswer": {
            "type": "object",
            "properties": {
                "quiz_id": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "total_correct_answer": {
                    "type": "integer"
                },
                "user_answer": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "helper.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "helper.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}